Modularize functional.py into a functional/ package
Split the 1,323-line numpyro_forecast/functional.py into submodules by
scope: models (Horizon + model-building primitives), svi (optimizer/guide
resolution + fit_svi), mcmc (kernel resolution + fit_mcmc), posterior
(draw_posterior singledispatch), prediction (forecast/predict_in_sample),
and a private _validation helper module. All public names are re-exported
from functional/__init__.py, so existing imports keep working; in-repo
private imports now target the defining modules.

great-docs.yml reference entries move to functional.<submodule>.<name>
paths (enforced by tests/test_docs_reference.py), docstring cross-
references point at the defining modules, and tests/test_functional.py is
split into six files mirroring the new layout (same 61 collected tests).

The _WARN_SKIP_PREFIXES first prefix now takes the package root via a
double dirname, since the constant moved one level down into
functional/mcmc.py.

Closes #48
